split quickcheck functionality across {core, parser}
we had a tricky problem with quickcheck because of Rust's limitation
on orphan instances. we want to be able to generate arbitrary `Expr`s,
however we want to condition `Name` struct arbitrary generation based on
constructs from the "parser" (concrete syntax). that means we are in a
pickle because we can't define orphan instances in the "parser". but we
have concrete syntax values which are in the parser and are thus not
accessible to `rep_lang_core`.

my solution here is to factor the `arbitrary` trait impl apart into many
`arbitrary_*` functions. I stub out the `arbitrary` trait impl for
`Expr` with a panic. then, in the "parser", I define a wrapper struct,
`WrappedExpr` and define trait impls on that for arbitrary & shrink.
this allows me to pass in the reserved keywords from the "parser".

seems like a fairly tidy solution, despite semi-substantial LOCs of
changes.
