+++ rep_lang_core/src/lib.rs
@@ -1,1 +1,3 @@
 pub mod syntax;
+
+pub mod test_helpers;
