+++ rep_lang_parser/src/pretty.rs
@@ -6,106 +6,97 @@ use rep_lang_core::syntax::{Defn, Expr, Expr::*, Lit, Lit::*, Name, PrimOp, Prim
 use crate::sp;
 use crate::util::pretty::parens;
 
-impl Expr {
-    pub fn ppr(&self) -> RcDoc<()> {
-        match &*self {
-            Var(name) => name.ppr(),
-            App(fun, arg) => {
-                let fun_ = fun.ppr();
-                let arg_ = arg.ppr();
-                parens(fun_.append(sp!()).append(arg_))
-            }
-            Lam(nm, bd) => {
-                let nm_ = nm.ppr();
-                let bd_ = bd.ppr();
-                parens(
-                    RcDoc::text("lam [")
-                        .append(nm_)
-                        .append(RcDoc::text("] "))
-                        .append(bd_),
-                )
-            }
-            Let(nm, e, bd) => {
-                let nm_ = nm.ppr();
-                let e_ = e.ppr();
-                let bd_ = bd.ppr();
-                parens(
-                    RcDoc::text("let ([")
-                        .append(nm_)
-                        .append(sp!())
-                        .append(e_)
-                        .append(RcDoc::text("]) "))
-                        .append(bd_),
-                )
-            }
-            Lit(x) => x.ppr(),
-            If(tst, thn, els) => {
-                let docs = vec![RcDoc::text("if"), tst.ppr(), thn.ppr(), els.ppr()];
-                parens(RcDoc::intersperse(docs, sp!()))
-            }
-            Fix(x) => parens(RcDoc::text("fix ").append(x.ppr())),
-            Prim(op) => op.ppr(),
+pub fn ppr_expr(expr: &Expr) -> RcDoc<()> {
+    match &*expr {
+        Var(name) => ppr_name(name),
+        App(fun, arg) => {
+            let fun_ = ppr_expr(fun);
+            let arg_ = ppr_expr(arg);
+            parens(fun_.append(sp!()).append(arg_))
         }
-    }
-}
-
-impl Lit {
-    pub fn ppr(&self) -> RcDoc<()> {
-        match *self {
-            LInt(i) => RcDoc::as_string(i),
-            LBool(true) => RcDoc::text("true"),
-            LBool(false) => RcDoc::text("false"),
+        Lam(nm, bd) => {
+            let nm_ = ppr_name(nm);
+            let bd_ = ppr_expr(bd);
+            parens(
+                RcDoc::text("lam [")
+                    .append(nm_)
+                    .append(RcDoc::text("] "))
+                    .append(bd_),
+            )
         }
-    }
-}
-
-impl PrimOp {
-    pub fn ppr(&self) -> RcDoc<()> {
-        match *self {
-            Add => RcDoc::text("+"),
-            Sub => RcDoc::text("-"),
-            Mul => RcDoc::text("*"),
-            Eql => RcDoc::text("=="),
-            Null => RcDoc::text("null"),
-            Map => RcDoc::text("map"),
-            Foldl => RcDoc::text("foldl"),
-            Pair => RcDoc::text("pair"),
-            Fst => RcDoc::text("fst"),
-            Snd => RcDoc::text("snd"),
-            Cons => RcDoc::text("cons"),
-            Nil => RcDoc::text("nil"),
-        }
-    }
-}
-
-impl Name {
-    pub fn ppr(&self) -> RcDoc<()> {
-        match &*self {
-            Name(s) => RcDoc::text(s),
-        }
-    }
-}
-
-impl Defn {
-    pub fn ppr(&self) -> RcDoc<()> {
-        match &*self {
-            Defn(nm, bd) => parens(
-                RcDoc::text("defn ")
-                    .append(nm.ppr())
+        Let(nm, e, bd) => {
+            let nm_ = ppr_name(nm);
+            let e_ = ppr_expr(e);
+            let bd_ = ppr_expr(bd);
+            parens(
+                RcDoc::text("let ([")
+                    .append(nm_)
                     .append(sp!())
-                    .append(bd.ppr()),
-            ),
+                    .append(e_)
+                    .append(RcDoc::text("]) "))
+                    .append(bd_),
+            )
         }
+        Lit(x) => ppr_lit(x),
+        If(tst, thn, els) => {
+            let tst_ = ppr_expr(tst);
+            let thn_ = ppr_expr(thn);
+            let els_ = ppr_expr(els);
+            let docs = vec![RcDoc::text("if"), tst_, thn_, els_];
+            parens(RcDoc::intersperse(docs, sp!()))
+        }
+        Fix(x) => parens(RcDoc::text("fix ").append(ppr_expr(x))),
+        Prim(op) => ppr_primop(op),
     }
 }
 
-impl Program {
-    pub fn ppr(&self) -> RcDoc<()> {
-        let docs = self
-            .p_defns
-            .iter()
-            .map(|d| d.ppr())
-            .chain(iter::once(self.p_body.ppr()));
-        RcDoc::intersperse(docs, "\n\n")
+pub fn ppr_lit(lit: &Lit) -> RcDoc<()> {
+    match *lit {
+        LInt(i) => RcDoc::as_string(i),
+        LBool(true) => RcDoc::text("true"),
+        LBool(false) => RcDoc::text("false"),
     }
+}
+
+pub fn ppr_primop(op: &PrimOp) -> RcDoc<()> {
+    match *op {
+        Add => RcDoc::text("+"),
+        Sub => RcDoc::text("-"),
+        Mul => RcDoc::text("*"),
+        Eql => RcDoc::text("=="),
+        Null => RcDoc::text("null"),
+        Map => RcDoc::text("map"),
+        Foldl => RcDoc::text("foldl"),
+        Pair => RcDoc::text("pair"),
+        Fst => RcDoc::text("fst"),
+        Snd => RcDoc::text("snd"),
+        Cons => RcDoc::text("cons"),
+        Nil => RcDoc::text("nil"),
+    }
+}
+
+pub fn ppr_name(name: &Name) -> RcDoc<()> {
+    match &*name {
+        Name(s) => RcDoc::text(s),
+    }
+}
+
+pub fn ppr_defn(defn: &Defn) -> RcDoc<()> {
+    match &*defn {
+        Defn(nm, bd) => parens(
+            RcDoc::text("defn ")
+                .append(ppr_name(nm))
+                .append(sp!())
+                .append(ppr_expr(bd)),
+        ),
+    }
+}
+
+pub fn ppr_program(prog: &Program) -> RcDoc<()> {
+    let docs = prog
+        .p_defns
+        .iter()
+        .map(|d| ppr_defn(d))
+        .chain(iter::once(ppr_expr(&prog.p_body)));
+    RcDoc::intersperse(docs, "\n\n")
 }
